Added support for gradle in addition to maven. For those of us allergic to xml and well maven ;-)
This initial support was created by running gradle init --pom which uses the new (incubating) build-init plugin for gradle. The generated gradle file did include an invalid "system" group. Not sure if this is a bug in the plugin, some fancy maven feature that isn't supported or something else entirely. It's currently commented out and will need to be reviewed.

An additional enhancement to this will include running the server through the application plugin. That will be added once I figure out exactly how this grand vision of Eirik's will unfold.